fix: allow tagging the same commit with multiple tags
When using prefixes the same commit might have different tags.

The jgit code did not take the prefix into account and stored tags in a map:
commit → tag

If multiple tags with different prefixes referenced the same commit only the last
tag was stored in this map.

This commit takes a prefix into account and ignores all tags which do not start
with the given prefix.